Fix count_distinct aggregations and reorganize critical tests
- Fix count_distinct metrics without sql field by using primary key instead of metric name
- Reorganize test_critical_fixes into appropriate test modules:
  - Filter tests (HAVING vs WHERE) → test_filters.py
  - Column prefix tests → test_basic.py
  - Count distinct tests → test_basic.py
  - DuckDB path tests → test_duckdb_adapter.py
  - Query parameters test → test_parameters.py
