+++ tests/metrics/test_filters.py
@@ -1,6 +1,9 @@
 """Test metric-level filters."""
 
-from sidemantic import Dimension, Metric, Model
+import duckdb
+
+from sidemantic import Dimension, Metric, Model, SemanticLayer
+from tests.utils import df_rows
 
 
 def test_metric_level_filter_basic(layer):
@@ -243,3 +246,185 @@ def test_metric_filter_multiple_columns_not_in_dimensions(layer):
     cte_match = sql.split("FROM")[0]
     assert "status" in cte_match, f"'status' should be in CTE SELECT. CTE: {cte_match}"
     assert "payment_method" in cte_match, f"'payment_method' should be in CTE SELECT. CTE: {cte_match}"
+
+
+def test_query_level_metric_filters_use_having(layer):
+    """Test that query-level filters on metrics use HAVING clause, not WHERE.
+
+    Bug: Query filters like orders.revenue > 100 were applied in WHERE clause
+    against _raw columns before aggregation, giving wrong results.
+
+    Fix: Detect when filter references a metric and use HAVING instead.
+    """
+    orders = Model(
+        name="orders",
+        table="orders_table",
+        primary_key="order_id",
+        dimensions=[
+            Dimension(name="region", type="categorical"),
+        ],
+        metrics=[
+            Metric(name="revenue", agg="sum", sql="amount"),
+        ],
+    )
+
+    layer.add_model(orders)
+
+    # Filter on aggregated metric should use HAVING
+    sql = layer.compile(metrics=["orders.revenue"], dimensions=["orders.region"], filters=["orders.revenue > 100"])
+
+    # Should have HAVING clause, not WHERE clause with revenue_raw
+    assert "HAVING" in sql
+    assert "revenue > 100" in sql
+    # Should NOT filter on raw column before aggregation
+    assert "revenue_raw > 100" not in sql
+
+
+def test_dimension_filters_use_where(layer):
+    """Test that filters on dimensions still use WHERE clause."""
+    layer = SemanticLayer()
+
+    orders = Model(
+        name="orders",
+        table="orders_table",
+        primary_key="order_id",
+        dimensions=[
+            Dimension(name="region", type="categorical"),
+            Dimension(name="status", type="categorical"),
+        ],
+        metrics=[
+            Metric(name="revenue", agg="sum", sql="amount"),
+        ],
+    )
+
+    layer.add_model(orders)
+
+    # Filter on dimension should use WHERE
+    sql = layer.compile(
+        metrics=["orders.revenue"], dimensions=["orders.region"], filters=["orders.status = 'completed'"]
+    )
+
+    # Should have WHERE clause
+    assert "WHERE" in sql
+    assert "status = 'completed'" in sql
+    # Should NOT have HAVING for dimension filter
+    assert "HAVING" not in sql
+
+
+def test_mixed_filters_separate_where_and_having(layer):
+    """Test that mixed metric and dimension filters use both WHERE and HAVING."""
+    layer = SemanticLayer()
+
+    orders = Model(
+        name="orders",
+        table="orders_table",
+        primary_key="order_id",
+        dimensions=[
+            Dimension(name="region", type="categorical"),
+            Dimension(name="status", type="categorical"),
+        ],
+        metrics=[
+            Metric(name="revenue", agg="sum", sql="amount"),
+        ],
+    )
+
+    layer.add_model(orders)
+
+    sql = layer.compile(
+        metrics=["orders.revenue"],
+        dimensions=["orders.region"],
+        filters=["orders.status = 'completed'", "orders.revenue > 100"],
+    )
+
+    # Should have both WHERE (for dimension) and HAVING (for metric)
+    assert "WHERE" in sql
+    assert "status = 'completed'" in sql
+    assert "HAVING" in sql
+    assert "revenue > 100" in sql
+
+
+def test_metric_level_filters_still_use_where(layer):
+    """Test that Metric.filters (row-level filters) still use WHERE clause.
+
+    Metric.filters are row-level filters that should filter before aggregation.
+    They're different from query-level filters on aggregated metrics.
+    """
+    layer = SemanticLayer()
+
+    orders = Model(
+        name="orders",
+        table="orders_table",
+        primary_key="order_id",
+        dimensions=[
+            Dimension(name="region", type="categorical"),
+        ],
+        metrics=[
+            Metric(name="completed_revenue", agg="sum", sql="amount", filters=["{model}.status = 'completed'"]),
+        ],
+    )
+
+    layer.add_model(orders)
+
+    sql = layer.compile(metrics=["orders.completed_revenue"], dimensions=["orders.region"])
+
+    # Metric-level filter should be in WHERE clause (applied to rows before aggregation)
+    assert "WHERE" in sql
+    assert "status = 'completed'" in sql
+
+
+def test_having_filter_with_actual_data(layer):
+    """Integration test with real DuckDB data to verify HAVING filter correctness."""
+    conn = duckdb.connect(":memory:")
+
+    # Create test data
+    conn.execute("""
+        CREATE TABLE orders (
+            order_id INTEGER,
+            customer_id INTEGER,
+            region VARCHAR,
+            status VARCHAR,
+            amount DECIMAL(10, 2)
+        )
+    """)
+
+    conn.execute("""
+        INSERT INTO orders VALUES
+        (1, 101, 'US', 'completed', 50),
+        (2, 101, 'US', 'completed', 150),  -- customer 101: 200 total
+        (3, 102, 'EU', 'completed', 300),
+        (4, 102, 'EU', 'pending', 75),      -- EU total: 375, completed: 300
+        (5, 103, 'US', 'cancelled', 25)
+    """)
+
+    layer = SemanticLayer()
+
+    orders = Model(
+        name="orders",
+        table="orders",
+        primary_key="order_id",
+        dimensions=[
+            Dimension(name="region", type="categorical"),
+        ],
+        metrics=[
+            Metric(name="revenue", agg="sum", sql="amount"),
+        ],
+    )
+
+    layer.conn = conn
+    layer.add_model(orders)
+
+    # Test HAVING filter (should filter aggregated revenue)
+    result = layer.query(
+        metrics=["orders.revenue"],
+        dimensions=["orders.region"],
+        filters=["orders.revenue >= 200"],  # Should use HAVING
+    )
+    rows = df_rows(result)
+
+    # Should only return regions with total revenue >= 200
+    # US: 50+150+25 = 225, EU: 300+75 = 375
+    # Both should be included
+    assert len(rows) == 2
+    revenues = {row[0]: row[1] for row in rows}
+    assert revenues["US"] == 225.0
+    assert revenues["EU"] == 375.0
